feat: add changelog_ignore_body customization
Setting changelog_ignore_body to true will ignore the body of the commit
message when processing commits for the changelog.

This mostly poses an issue for breaking changes, which may provide additional
details in the body. With this flag set to true, only the subject of the commit
message is processed.